feat: #75 Fix sorting strategy
Also rename  Sorter into SortingStrategy and introduce Sorter to represent the step as others (Filter, Pruner...)
