+++ modules/core/src/main/kotlin/datamaintain/core/Datamaintain.kt
@@ -2,12 +2,11 @@ package datamaintain.core
 
 import datamaintain.core.config.DatamaintainConfig
 import datamaintain.core.report.Report
-import datamaintain.core.script.Script
 import datamaintain.core.step.Filter
 import datamaintain.core.step.Pruner
 import datamaintain.core.step.Scanner
 import datamaintain.core.step.executor.Executor
-import datamaintain.core.step.sort.ByLengthAndCaseInsensitiveAlphabeticalSorter
+import datamaintain.core.step.sort.Sorter
 import mu.KotlinLogging
 
 private val logger = KotlinLogging.logger {}
@@ -29,10 +28,7 @@ class Datamaintain(config: DatamaintainConfig) {
     fun updateDatabase(): Report {
         return Scanner(context).scan()
                 .let { scripts -> Filter(context).filter(scripts) }
-                .let { scripts ->
-                    ByLengthAndCaseInsensitiveAlphabeticalSorter(context.config)
-                            .sort(scripts, Script::identifier)
-                }
+                .let { scripts -> Sorter(context).sort(scripts) }
                 .let { scripts -> Pruner(context).prune(scripts) }
                 .let { scripts -> Executor(context).execute(scripts) }
     }
