+++ modules/core/src/main/kotlin/datamaintain/core/step/sort/Sorter.kt
@@ -1,8 +1,22 @@
 package datamaintain.core.step.sort
 
-import datamaintain.core.config.DatamaintainConfig
+import datamaintain.core.Context
 import datamaintain.core.script.Script
+import datamaintain.core.script.ScriptWithContent
+import mu.KotlinLogging
 
-abstract class Sorter<U>(protected val config: DatamaintainConfig) {
-    abstract fun <T : Script> sort(scripts: List<T>, getter: (T) -> U): List<T>
+private val logger = KotlinLogging.logger {}
+
+class Sorter(private val context: Context) {
+    fun sort(scripts: List<ScriptWithContent>): List<ScriptWithContent> {
+        logger.info { "Sort scripts..." }
+
+        val sortingStrategy = ByCaseInsensitiveSeparatorFreeAlphabeticalSortingStrategy();
+        val sortedScripts = sortingStrategy.sort(scripts, Script::identifier)
+
+        logger.info { "Scripts sorted" }
+        logger.info { "" }
+
+        return sortedScripts
+    }
 }
