+++ modules/core/src/main/kotlin/datamaintain/core/step/sort/SortingStrategy.kt
@@ -1,0 +1,7 @@
+package datamaintain.core.step.sort
+
+import datamaintain.core.script.Script
+
+abstract class SortingStrategy<U>() {
+    abstract fun <T : Script> sort(scripts: List<T>, getter: (T) -> U): List<T>
+}
